fix: 修复5个Linux编译错误
- platform/linux.rs: log_error! 加 crate::前缀; check_file_capabilities 加 pub
- Cargo.toml: libc 改为 linux+macos unconditional 依赖
- tun/linux.rs: mtu u32→u16 转换
- proxy/https_tunnel.rs: 避开hyper::rt::Read/Write trait, 改用split+join!手动双向复制

+++ src-tauri/src/tun/linux.rs
@@ -13,7 +13,7 @@ impl TunManager for LinuxTunManager {
         let mut tun_cfg = tun_easytier::Configuration::default();
         tun_cfg
             .layer(tun_easytier::Layer::L3)
-            .mtu(config.mtu)
+            .mtu(config.mtu as u16)
             .up();
         if let Some(ref name) = config.dev_name {
             if !name.is_empty() {
